+++ src/models/ground_penalty.js
